pyproject.toml
[project]
name = "dbutils"
version = "0.1.0"
description = "DB2 relationship resolver"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "rich>=13.0.0",
    "textual>=6.6.0",
    "PySide6>=6.5.0",
    "JPype1>=1.5.0",
    "JayDeBeApi>=1.2.3",
]

[build-system]
requires = ["setuptools>=61", "wheel", "Cython>=3.0.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
  gui = [
      "PySide6>=6.5.0",
      "PyQt6>=6.5.0"
  ]
  dev = [
      "ruff>=0.1.0",
      "pytest>=7.0.0",
      "PySide6>=6.5.0",
      "PyQt6>=6.5.0",
      "JPype1>=1.5.0",
      "JayDeBeApi>=1.2.3"
  ]
  cython = [
      "cython>=3.0.0",
      "numpy>=1.21.0"
  ]

[project.scripts]
  db-relate = "dbutils.db_relate:main"
  db-map = "dbutils.map_db:main"
  db-analyze = "dbutils.db_analyze:main"
  db-search = "dbutils.db_search:main"
  db-diff = "dbutils.db_diff:main"
  db-health = "dbutils.db_health:main"
  db-browser = "dbutils.main_launcher:main"
  db-browser-gui = "dbutils.gui.qt_app:main"
  db-table-sizes = "dbutils.db_table_sizes:main"
  db-indexes = "dbutils.db_indexes:main"
  db-inferred-ref-coverage = "dbutils.db_inferred_ref_coverage:main"
  db-inferred-orphans = "dbutils.db_inferred_orphans:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = {find = {where = ["src"], include = ["dbutils*"]}}

[tool.uv]
package = true

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
]

[tool.ruff.lint.per-file-ignores]
"src/dbutils/__init__.py" = ["F401"]
"src/dbutils/db_relate.py" = [
    "E501",  # Allow long lines for SQL queries
    "C901"   # Allow complex function for relationship inference
]
"src/dbutils/map_db.py" = [
    "E501",  # Allow long lines for SQL queries
    "C901"   # Allow complex function for relationship inference
]
"src/dbutils/db_diff.py" = [
    "C901"   # Allow complex function for table comparison
]
"src/dbutils/db_health.py" = [
    "C901"   # Allow complex functions for health checks
]
"src/dbutils/db_search.py" = [
    "E501"   # Allow long lines for SQL queries
]
"src/dbutils/db_browser.py" = [
    "I001",   # Allow imports in try blocks without sorting
    "C901",   # Allow complex functions for search functionality
    "E501"    # Allow long lines for mock data definitions
]

